feat(ci): switch vault secret loading to nsc vault export
Replaces the bespoke scripts/load-ns-secrets.mjs (VaultService/DescribeObject via NSC_VAULT_TOKEN) with the namespace team's simpler recipe: 'nsc vault export --envdef /tmp/vault.envdef --shell=bash | eval' under OIDC runner auth.

- Each runs-on: namespace-profile-default now passes permissions.additional_grant=vault/object:*:describe so jobs can read vault objects with no static token.

- New scripts/write-vault-envdef.mjs reuses scripts/provider-vars.json to enumerate only the per-job subset, so per-provider scoping is preserved.

- scripts/load-ns-secrets.mjs deleted; the old NSC_VAULT_TOKEN repo secret is unreferenced and safe to remove once nsc-vault-federation lands.

- Touches 7 benchmark workflows (browser, browser-throughput, sandbox-dax, sandbox-tti, snapshot-fork, storage, ai-gateway); scale-image.yml is unchanged (uses NSC_TOKEN for nsc docker login, not vault).
