--- scripts/load-ns-secrets.mjs
@@ -1,252 +1,0 @@
-#!/usr/bin/env node
-// Resolve secrets from the Namespace vault into the GitHub Actions environment.
-// Namespace vault secret resolver for GitHub Actions.
-//
-// Each vault secret carries a `name=<ENV_VAR>` label.
-// This script lists the vault, reveals the requested secrets via the
-// VaultService/DescribeObject API, masks them, and appends them to GITHUB_ENV.
-//
-// The Namespace CLI has NO secret read-back command, so revealing goes through
-// the Connect API directly:
-//   POST https://us.compute.namespaceapis.com/namespace.cloud.vault.v1beta.VaultService/{ListObjects,DescribeObject}
-// authenticated with a bearer token (NSC_VAULT_TOKEN) that holds
-//   {resource_type:"vault/object", resource_id:"*", actions:["list","describe"]}.
-//
-// Usage:
-//   node scripts/load-ns-secrets.mjs --mode sandbox --provider e2b  # SCOPED: just this provider's declared vars (CI default)
-//   node scripts/load-ns-secrets.mjs AWS_REGION E2B_API_KEY  # only these explicit vars
-//   node scripts/load-ns-secrets.mjs --section aws-s3        # only that section label
-//   node scripts/load-ns-secrets.mjs                         # all name-labeled secrets (avoid in matrix CI: request burst)
-//   node scripts/load-ns-secrets.mjs --print ...             # local check: reveals secrets, prints only names+byte-lengths (never the values), skips GITHUB_ENV
-//
-// --mode/--provider look up scripts/provider-vars.json (generated by
-// gen-provider-vars.ts from each provider's requiredEnvVars). --mode defaults
-// to "sandbox". A required var absent from the vault is a warning, not a failure.
-//
-// Env:
-//   NSC_VAULT_TOKEN     (always required, including --print — it still calls the vault API) bearer token
-//   NSC_VAULT_ENDPOINT  override API host (default us.compute.namespaceapis.com)
-//   GITHUB_ENV          set by Actions; required unless --print
-import { appendFileSync, readFileSync } from 'node:fs';
-import { request } from 'node:https';
-import { randomUUID } from 'node:crypto';
-
-const argv = process.argv.slice(2);
-const printOnly = argv.includes('--print');
-
-// --flag <value> options; the value index is tracked so it isn't mistaken
-// for a positional var name.
-const valueIdxs = new Set();
-function flagValue(name) {
-  const i = argv.indexOf(name);
-  if (i === -1) return null;
-  valueIdxs.add(i + 1);
-  return argv[i + 1] ?? null;
-}
-const section = flagValue('--section');
-const mode = flagValue('--mode');
-const provider = flagValue('--provider');
-// Positional args = explicit var names (excludes flags and their values).
-const wantNames = argv.filter((a, i) => !a.startsWith('--') && !valueIdxs.has(i));
-
-const token = process.env.NSC_VAULT_TOKEN;
-const host = process.env.NSC_VAULT_ENDPOINT || 'us.compute.namespaceapis.com';
-const githubEnv = process.env.GITHUB_ENV;
-const CONCURRENCY = 8;
-
-if (!token) {
-  console.error('NSC_VAULT_TOKEN is not set');
-  process.exit(1);
-}
-if (!printOnly && !githubEnv) {
-  console.error('GITHUB_ENV is not set — run inside GitHub Actions, or pass --print to dry-run');
-  process.exit(1);
-}
-
-const SVC = 'namespace.cloud.vault.v1beta.VaultService';
-
-// One Connect-JSON unary call, with retry/backoff on transient failures.
-function call(method, body) {
-  const payload = JSON.stringify(body);
-  const attempt = (n) =>
-    new Promise((resolve, reject) => {
-      const req = request(
-        `https://${host}/${SVC}/${method}`,
-        {
-          method: 'POST',
-          headers: {
-            Authorization: `Bearer ${token}`,
-            'Content-Type': 'application/json',
-            'Content-Length': Buffer.byteLength(payload),
-          },
-        },
-        (res) => {
-          let data = '';
-          res.on('data', (c) => (data += c));
-          res.on('end', () => {
-            if (res.statusCode === 200) return resolve(JSON.parse(data || '{}'));
-            const retryable = res.statusCode >= 500 || res.statusCode === 429;
-            const err = new Error(`${method} HTTP ${res.statusCode}: ${data.slice(0, 300)}`);
-            err.retryable = retryable;
-            reject(err);
-          });
-        },
-      );
-      req.on('error', (e) => {
-        e.retryable = true;
-        reject(e);
-      });
-      req.write(payload);
-      req.end();
-    });
-
-  return (async () => {
-    const maxAttempts = 5;
-    for (let n = 1; n <= maxAttempts; n++) {
-      try {
-        return await attempt(n);
-      } catch (e) {
-        if (!e.retryable || n === maxAttempts) throw e;
-        const wait = Math.min(15000, 500 * 2 ** (n - 1));
-        console.error(`${method} failed (attempt ${n}/${maxAttempts}): ${e.message}; retrying in ${wait}ms`);
-        await new Promise((r) => setTimeout(r, wait));
-      }
-    }
-  })();
-}
-
-function labelOf(obj, key) {
-  return (obj.labels || []).find((l) => l.name === key)?.value;
-}
-
-// Mask every line so multiline secrets stay hidden in logs.
-function exportVar(name, value) {
-  for (const line of String(value).split(/\r?\n/)) {
-    if (line.trim().length) console.log(`::add-mask::${line}`);
-  }
-  const delim = `NS_EOF_${randomUUID().replace(/-/g, '')}`;
-  appendFileSync(githubEnv, `${name}<<${delim}\n${value}\n${delim}\n`);
-}
-
-async function pool(items, worker, size) {
-  const results = new Array(items.length);
-  let i = 0;
-  const runners = Array.from({ length: Math.min(size, items.length) }, async () => {
-    while (i < items.length) {
-      const idx = i++;
-      results[idx] = await worker(items[idx], idx);
-    }
-  });
-  await Promise.all(runners);
-  return results;
-}
-
-(async () => {
-  const { objects = [] } = await call('ListObjects', {});
-  // A secret's env-var name comes from its name= label (set by env-to-vault.mjs).
-  // Fall back to the (trimmed) description when it looks like an env var name, so
-  // secrets added manually via `nsc vault add --description FOO_API_KEY` (no label)
-  // are still picked up. The Connect JSON API returns camelCase `objectId`; the
-  // nsc CLI uses snake_case `object_id` — accept either.
-  const ENV_NAME = /^[A-Z_][A-Z0-9_]*$/;
-  // De-dupe by name: a labeled secret always wins over a description-derived one;
-  // otherwise pick the lowest object id so the result is deterministic (ListObjects
-  // order is not). Without this, two secrets mapping to the same name would both be
-  // exported and whichever came last would silently win.
-  const byName = new Map();
-  for (const o of objects) {
-    const id = o.objectId ?? o.object_id;
-    if (!id) continue;
-    const nameLabel = labelOf(o, 'name');
-    const desc = (o.description || '').trim();
-    const labeled = Boolean(nameLabel);
-    const name = nameLabel || (ENV_NAME.test(desc) ? desc : undefined);
-    if (!name) continue;
-    const entry = { id, name, section: labelOf(o, 'section'), labeled };
-    const prev = byName.get(name);
-    if (!prev) { byName.set(name, entry); continue; }
-    const better = labeled !== prev.labeled ? labeled : id < prev.id;
-    console.error(`WARNING: multiple vault secrets map to ${name}; using ${better ? id : prev.id} (labeled wins, else lowest id)`);
-    if (better) byName.set(name, entry);
-  }
-  let targets = [...byName.values()];
-
-  // --provider (+ --mode): load only that provider's declared requiredEnvVars,
-  // read from the generated manifest. This is the scoped path used by CI so a
-  // job reveals ~a handful of secrets instead of all 81 (avoids a request burst).
-  if (provider) {
-    const manPath = new URL('./provider-vars.json', import.meta.url);
-    let manifest;
-    try {
-      manifest = JSON.parse(readFileSync(manPath, 'utf8'));
-    } catch {
-      console.error('provider-vars.json missing — run: npx tsx scripts/gen-provider-vars.ts');
-      process.exit(1);
-    }
-    const modeKey = mode || 'sandbox';
-    const modeMap = manifest[modeKey];
-    if (!modeMap) {
-      console.error(`Unknown --mode "${modeKey}". Known: ${Object.keys(manifest).join(', ')}`);
-      process.exit(1);
-    }
-    const vars = modeMap[provider];
-    if (!vars) {
-      console.error(`Provider "${provider}" not found in mode "${modeKey}". Known: ${Object.keys(modeMap).join(', ')}`);
-      process.exit(1);
-    }
-    const found = new Set(targets.map((o) => o.name));
-    // A required var absent from the vault is a WARNING, not a failure: some
-    // requiredEnvVars (e.g. ARCHIL_DISK_ID, MODAL_TOKEN_ID) are not in .env/the
-    // vault today, and the benchmark itself surfaces a truly-missing cred.
-    const missing = vars.filter((v) => !found.has(v));
-    if (missing.length) {
-      console.error(`WARNING: ${modeKey}/${provider} requires vars not in vault (skipping): ${missing.join(', ')}`);
-    }
-    const want = new Set(vars);
-    targets = targets.filter((o) => want.has(o.name));
-  } else if (section) {
-    targets = targets.filter((o) => o.section === section);
-  } else if (wantNames.length) {
-    const want = new Set(wantNames);
-    const found = new Set(targets.map((o) => o.name));
-    const missing = wantNames.filter((n) => !found.has(n));
-    if (missing.length) {
-      console.error(`Requested vars not found in vault: ${missing.join(', ')}`);
-      process.exit(1);
-    }
-    targets = targets.filter((o) => want.has(o.name));
-  }
-
-  if (!targets.length) {
-    // In provider mode an empty set means this provider's creds aren't in the
-    // vault; mirror 1Password's loader (load 0, continue) and let the benchmark
-    // report the missing cred. In other modes an empty set is a real error.
-    if (provider) {
-      console.error(`No vault secrets available for ${mode || 'sandbox'}/${provider}; continuing.`);
-      process.exit(0);
-    }
-    console.error('No matching secrets to load.');
-    process.exit(1);
-  }
-
-  const revealed = await pool(
-    targets,
-    async (t) => {
-      const { value } = await call('DescribeObject', { object_id: t.id });
-      return { ...t, value };
-    },
-    CONCURRENCY,
-  );
-
-  for (const r of revealed) {
-    if (printOnly) {
-      console.log(`  ${r.name.padEnd(38)} section=${r.section ?? '-'} (${Buffer.byteLength(r.value ?? '', 'utf8')}B)`);
-    } else {
-      exportVar(r.name, r.value);
-    }
-  }
-  console.error(`Loaded ${revealed.length} secret(s) from Namespace vault${section ? ` (section=${section})` : ''}.`);
-})().catch((e) => {
-  console.error(e.message || e);
-  process.exit(1);
-});
