+++ scripts/write-vault-envdef.mjs
@@ -1,0 +1,60 @@
+#!/usr/bin/env node
+// Emit a Namespace vault --envdef file (one `KEY=sec_<id>` line per secret)
+// for the given --mode/--provider pair, or for explicit positional names.
+//
+// Mirrors scripts/provider-vars.json (generated by gen-provider-vars.ts) so
+// per-job secret scoping stays the same — only the loader changes.
+//
+// Usage:
+//   node scripts/write-vault-envdef.mjs --mode browser --provider hyperbrowser
+//   node scripts/write-vault-envdef.mjs --mode sandbox --provider modals
+//   node scripts/write-vault-envdef.mjs INGEST_URL INGEST_SECRET   # explicit names
+//   node scripts/write-vault-envdef.mjs                          # all known names
+//
+// Consumed by:
+//   eval "$(nsc vault export --envdef /tmp/vault.envdef --shell=bash)"
+// Each `sec_<id>` value in the envdef is a placeholder; `nsc vault export`
+// resolves each line against the vault (label or id) and exports the revealed
+// value into the bash environment under `KEY`.
+import { readFileSync } from 'node:fs';
+
+const PLACEHOLDER = 'sec_dummy';
+
+const argv = process.argv.slice(2);
+const valueIdxs = new Set();
+function flagValue(name) {
+  const i = argv.indexOf(name);
+  if (i === -1) return null;
+  valueIdxs.add(i + 1);
+  return argv[i + 1] ?? null;
+}
+const mode = flagValue('--mode');
+const provider = flagValue('--provider');
+const positional = argv.filter((a, i) => !a.startsWith('--') && !valueIdxs.has(i));
+
+let names;
+if (provider) {
+  const manifest = JSON.parse(readFileSync(new URL('./provider-vars.json', import.meta.url), 'utf8'));
+  const modeKey = mode || 'sandbox';
+  const vars = manifest[modeKey]?.[provider];
+  if (!vars) {
+    console.error(`No manifest entries for ${modeKey}/${provider}. Known modes: ${Object.keys(manifest).join(', ')}`);
+    process.exit(1);
+  }
+  names = vars;
+} else if (positional.length) {
+  names = positional;
+} else {
+  // No scope: every name across every manifest mode (used by the single-job
+  // AI gateway workflow, which round-robins all gateways in one process).
+  const manifest = JSON.parse(readFileSync(new URL('./provider-vars.json', import.meta.url), 'utf8'));
+  const seen = new Set();
+  for (const modeMap of Object.values(manifest)) {
+    for (const vars of Object.values(modeMap)) {
+      for (const v of vars) seen.add(v);
+    }
+  }
+  names = [...seen].sort();
+}
+
+process.stdout.write(names.map((n) => `${n}=${PLACEHOLDER}\n`).join(''));
